Added Random Weight Guessing (RWG) to FANN model
NOTE: one test fails, the shared example on solving the XOR using
RWG. Several likely reasons, mostly the function being minimal in size
(and so harder to guess), and I am not controlling yet the range for
weight generation (looks like all weights are close to 0).
Note that the XOR being nonlinear should not pose a problem since it's
a nonlinear function we're guessing.